moved types into one .d.ts and created a barrel
