--- src/core/wo-dispatcher.core.d.ts
@@ -1,5 +1,0 @@
-import { WODispatcherConfiguration } from '../entities/wo-dispatcher-configuration.entity';
-import { WOCommandSet } from '../entities/wo-command-set.entity';
-export declare class WODispatcher {
-    static dispatch(configuration: WODispatcherConfiguration[], command_set: WOCommandSet): void;
-}
