--- src/core/wo-parser.core.d.ts
@@ -1,6 +1,0 @@
-import { WODispatcherConfiguration } from '../entities/wo-dispatcher-configuration.entity';
-export declare class WOParser {
-    private static command_set;
-    static startParser(dispatcher_conf: WODispatcherConfiguration[]): void;
-    private static resetCommandSet();
-}
