--- src/entities/wo-command-set.entity.d.ts
@@ -1,4 +1,0 @@
-export declare class WOCommandSet {
-    command: string | null;
-    flags: string[] | null;
-}
