--- src/entities/wo-dispatcher-configuration.entity.d.ts
@@ -1,6 +1,0 @@
-export declare class WODispatcherConfiguration {
-    command: string;
-    aliases?: string[];
-    flags?: string[];
-    action: (flags: string[]) => void;
-}
