--- src/entities/wo-sdk-configuration.entity.d.ts
@@ -1,10 +1,0 @@
-import { WODispatcherConfiguration } from './wo-dispatcher-configuration.entity';
-export declare class WOSDKConfiguration {
-    dispatcher: WODispatcherConfiguration[];
-    input_field: {
-        clear_after_submit: boolean;
-    };
-    general: {
-        theme: string;
-    };
-}
