--- src/enums/wo-severity.enum.d.ts
@@ -1,6 +1,0 @@
-export declare enum WOSeverityEnum {
-    message = 0,
-    error = 1,
-    warning = 2,
-    info = 3,
-}
